Separate jsonnet fmt into its own executable.
