+++ cmd/jsonnet.cpp
@@ -27,9 +27,10 @@ limitations under the License.
 #include <string>
 #include <vector>
 
+#include "utils.h"
+
 extern "C" {
 #include <libjsonnet.h>
-#include <libjsonnet_fmt.h>
 }
 
 #ifdef _WIN32
@@ -37,41 +38,6 @@ const char PATH_SEP = ';';
 #else
 const char PATH_SEP = ':';
 #endif
-
-std::string next_arg(unsigned &i, const std::vector<std::string> &args)
-{
-    i++;
-    if (i >= args.size()) {
-        std::cerr << "Expected another commandline argument." << std::endl;
-        exit(EXIT_FAILURE);
-    }
-    return args[i];
-}
-
-/** Collect commandline args into a vector of strings, and expand -foo to -f -o -o. */
-std::vector<std::string> simplify_args(int argc, const char **argv)
-{
-    std::vector<std::string> r;
-    for (int i = 1; i < argc; ++i) {
-        std::string arg = argv[i];
-        if (arg == "--") {
-            // Add this arg and all remaining ones without simplification.
-            r.push_back(arg);
-            while ((++i) < argc)
-                r.push_back(argv[i]);
-            break;
-        }
-        // Check if it is of the form -abc and convert to -a -b -c
-        if (arg.length() > 2 && arg[0] == '-' && arg[1] != '-') {
-            for (unsigned j = 1; j < arg.length(); ++j) {
-                r.push_back("-" + arg.substr(j, 1));
-            }
-        } else {
-            r.push_back(arg);
-        }
-    }
-    return r;
-}
 
 void version(std::ostream &o)
 {
@@ -82,15 +48,9 @@ void usage(std::ostream &o)
 {
     version(o);
     o << "\n";
-    o << "General commandline:\n";
-    o << "jsonnet [<cmd>] {<option>} { <filename> }\n";
-    o << "Note: <cmd> defaults to \"eval\"\n";
+    o << "jsonnet {<option>} <filename>\n";
     o << "\n";
-    o << "The eval command:\n";
-    o << "jsonnet eval {<option>} <filename>\n";
-    o << "Note: Only one filename is supported\n";
-    o << "\n";
-    o << "Available eval options:\n";
+    o << "Available options:\n";
     o << "  -h / --help             This message\n";
     o << "  -e / --exec             Treat filename as code\n";
     o << "  -J / --jpath <dir>      Specify an additional library search dir (right-most wins)\n";
@@ -124,27 +84,6 @@ void usage(std::ostream &o)
     o << "JSONNET_PATH=d:c:a:b jsonnet\n";
     o << "jsonnet -J b -J a -J c -J d\n";
     o << "\n";
-    o << "The fmt command:\n";
-    o << "jsonnet fmt {<option>} { <filename> }\n";
-    o << "Note: Some options do not support multiple filenames\n";
-    o << "\n";
-    o << "Available fmt options:\n";
-    o << "  -h / --help             This message\n";
-    o << "  -e / --exec             Treat filename as code\n";
-    o << "  -o / --output-file <file> Write to the output file rather than stdout\n";
-    o << "  -i / --in-place         Update the Jsonnet file(s) in place.\n";
-    o << "  --test                  Exit with failure if reformatting changed the file(s).\n";
-    o << "  -n / --indent <n>       Number of spaces to indent by (default 2, 0 means no change)\n";
-    o << "  --max-blank-lines <n>   Max vertical spacing, 0 means no change (default 2)\n";
-    o << "  --string-style <d|s|l>  Enforce double, single (default) quotes or 'leave'\n";
-    o << "  --comment-style <h|s|l> # (h), // (s)(default), or 'leave'; never changes she-bang\n";
-    o << "  --[no-]pretty-field-names Use syntax sugar for fields and indexing (on by default)\n";
-    o << "  --[no-]pad-arrays       [ 1, 2, 3 ] instead of [1, 2, 3]\n";
-    o << "  --[no-]pad-objects      { x: 1, y: 2 } instead of {x: 1, y: 2} (on by default)\n";
-    o << "  --[no-]sort-imports     Sorting of imports (on by default)\n";
-    o << "  --debug-desugaring      Unparse the desugared AST without executing it\n";
-    o << "  --version               Print version\n";
-    o << "\n";
     o << "In all cases:\n";
     o << "<filename> can be - (stdin)\n";
     o << "Multichar options are expanded e.g. -abc becomes -a -b -c.\n";
@@ -154,26 +93,8 @@ void usage(std::ostream &o)
     o << std::endl;
 }
 
-long strtol_check(const std::string &str)
-{
-    const char *arg = str.c_str();
-    char *ep;
-    long r = std::strtol(arg, &ep, 10);
-    if (*ep != '\0' || *arg == '\0') {
-        std::cerr << "ERROR: invalid integer \"" << arg << "\"\n" << std::endl;
-        exit(EXIT_FAILURE);
-    }
-    return r;
-}
-
-enum Command {
-    EVAL,
-    FMT,
-};
-
 /** Class for representing configuration read from command line flags.  */
 struct JsonnetConfig {
-    Command cmd;
     std::vector<std::string> inputFiles;
     std::string outputFile;
     bool filenameIsCode;
@@ -183,17 +104,10 @@ struct JsonnetConfig {
     bool evalStream;
     std::string evalMultiOutputDir;
 
-    // FMT flags
-    bool fmtInPlace;
-    bool fmtTest;
-
     JsonnetConfig()
-        : cmd(EVAL),
-          filenameIsCode(false),
+        : filenameIsCode(false),
           evalMulti(false),
-          evalStream(false),
-          fmtInPlace(false),
-          fmtTest(false)
+          evalStream(false)
     {
     }
 };
@@ -253,13 +167,6 @@ static ArgStatus process_args(int argc, const char **argv, JsonnetConfig *config
     std::vector<std::string> remaining_args;
 
     unsigned i = 0;
-    if (args.size() > 0 && args[i] == "fmt") {
-        config->cmd = FMT;
-        i++;
-    } else if (args.size() > 0 && args[i] == "eval") {
-        config->cmd = EVAL;
-        i++;
-    }
 
     for (; i < args.size(); ++i) {
         const std::string &arg = args[i];
@@ -283,218 +190,144 @@ static ArgStatus process_args(int argc, const char **argv, JsonnetConfig *config
             while ((++i) < args.size())
                 remaining_args.push_back(args[i]);
             break;
-        } else if (config->cmd == EVAL) {
-            if (arg == "-s" || arg == "--max-stack") {
-                long l = strtol_check(next_arg(i, args));
-                if (l < 1) {
-                    std::cerr << "ERROR: invalid --max-stack value: " << l << std::endl;
-                    return ARG_FAILURE;
-                }
-                jsonnet_max_stack(vm, l);
-            } else if (arg == "-J" || arg == "--jpath") {
-                std::string dir = next_arg(i, args);
-                if (dir.length() == 0) {
-                    std::cerr << "ERROR: -J argument was empty string" << std::endl;
-                    return ARG_FAILURE;
-                }
-                if (dir[dir.length() - 1] != '/') {
-                    dir += '/';
-                }
-                jsonnet_jpath_add(vm, dir.c_str());
-            } else if (arg == "-V" || arg == "--ext-str") {
-                std::string var, val;
-                if (!get_var_val(next_arg(i, args), var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_var(vm, var.c_str(), val.c_str());
-            } else if (arg == "-E" || arg == "--var" || arg == "--env") {
-                // TODO(dcunnin): Delete this in a future release.
-                std::cerr << "WARNING: jsonnet eval -E, --var and --env are deprecated,"
-                          << " please use -V or --ext-str." << std::endl;
-                std::string var, val;
-                if (!get_var_val(next_arg(i, args), var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_var(vm, var.c_str(), val.c_str());
-            } else if (arg == "--ext-str-file") {
-                std::string var, val;
-                if (!get_var_file(next_arg(i, args), "importstr", var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "-F" || arg == "--file") {
-                // TODO(dcunnin): Delete this in a future release.
-                std::cerr << "WARNING: jsonnet eval -F and --file are deprecated,"
-                          << " please use --ext-str-file." << std::endl;
-                std::string var, val;
-                if (!get_var_file(next_arg(i, args), "importstr", var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "--ext-code") {
-                std::string var, val;
-                if (!get_var_val(next_arg(i, args), var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "--code-var" || arg == "--code-env") {
-                // TODO(dcunnin): Delete this in a future release.
-                std::cerr << "WARNING: jsonnet eval --code-var and --code-env are deprecated,"
-                          << " please use --ext-code." << std::endl;
-                std::string var, val;
-                if (!get_var_val(next_arg(i, args), var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "--ext-code-file") {
-                std::string var, val;
-                if (!get_var_file(next_arg(i, args), "import", var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "--code-file") {
-                // TODO(dcunnin): Delete this in a future release.
-                std::cerr << "WARNING: jsonnet eval --code-file is deprecated,"
-                          << " please use --ext-code-file." << std::endl;
-                std::string var, val;
-                if (!get_var_file(next_arg(i, args), "import", var, val))
-                    return ARG_FAILURE;
-                jsonnet_ext_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "-A" || arg == "--tla-str") {
-                std::string var, val;
-                if (!get_var_val(next_arg(i, args), var, val))
-                    return ARG_FAILURE;
-                jsonnet_tla_var(vm, var.c_str(), val.c_str());
-            } else if (arg == "--tla-str-file") {
-                std::string var, val;
-                if (!get_var_file(next_arg(i, args), "importstr", var, val))
-                    return ARG_FAILURE;
-                jsonnet_tla_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "--tla-code") {
-                std::string var, val;
-                if (!get_var_val(next_arg(i, args), var, val))
-                    return ARG_FAILURE;
-                jsonnet_tla_code(vm, var.c_str(), val.c_str());
-            } else if (arg == "--tla-code-file") {
-                std::string var, val;
-                if (!get_var_file(next_arg(i, args), "import", var, val))
-                    return ARG_FAILURE;
-                jsonnet_tla_code(vm, var.c_str(), val.c_str());
-
-            } else if (arg == "--gc-min-objects") {
-                long l = strtol_check(next_arg(i, args));
-                if (l < 0) {
-                    std::cerr << "ERROR: invalid --gc-min-objects value: " << l << std::endl;
-                    return ARG_FAILURE;
-                }
-                jsonnet_gc_min_objects(vm, l);
-            } else if (arg == "-t" || arg == "--max-trace") {
-                long l = strtol_check(next_arg(i, args));
-                if (l < 0) {
-                    std::cerr << "ERROR: invalid --max-trace value: " << l << std::endl;
-                    return ARG_FAILURE;
-                }
-                jsonnet_max_trace(vm, l);
-            } else if (arg == "--gc-growth-trigger") {
-                const char *arg = next_arg(i, args).c_str();
-                char *ep;
-                double v = std::strtod(arg, &ep);
-                if (*ep != '\0' || *arg == '\0') {
-                    std::cerr << "ERROR: invalid number \"" << arg << "\"" << std::endl;
-                    return ARG_FAILURE;
-                }
-                if (v < 0) {
-                    std::cerr << "ERROR: invalid --gc-growth-trigger \"" << arg << "\""
-                              << std::endl;
-                    return ARG_FAILURE;
-                }
-                jsonnet_gc_growth_trigger(vm, v);
-            } else if (arg == "-m" || arg == "--multi") {
-                config->evalMulti = true;
-                std::string output_dir = next_arg(i, args);
-                if (output_dir.length() == 0) {
-                    std::cerr << "ERROR: -m argument was empty string" << std::endl;
-                    return ARG_FAILURE;
-                }
-                if (output_dir[output_dir.length() - 1] != '/') {
-                    output_dir += '/';
-                }
-                config->evalMultiOutputDir = output_dir;
-            } else if (arg == "-y" || arg == "--yaml-stream") {
-                config->evalStream = true;
-            } else if (arg == "-S" || arg == "--string") {
-                jsonnet_string_output(vm, 1);
-            } else if (arg.length() > 1 && arg[0] == '-') {
-                std::cerr << "ERROR: unrecognized argument: " << arg << std::endl;
+        } else if (arg == "-s" || arg == "--max-stack") {
+            long l = strtol_check(next_arg(i, args));
+            if (l < 1) {
+                std::cerr << "ERROR: invalid --max-stack value: " << l << std::endl;
                 return ARG_FAILURE;
-            } else {
-                remaining_args.push_back(args[i]);
             }
+            jsonnet_max_stack(vm, l);
+        } else if (arg == "-J" || arg == "--jpath") {
+            std::string dir = next_arg(i, args);
+            if (dir.length() == 0) {
+                std::cerr << "ERROR: -J argument was empty string" << std::endl;
+                return ARG_FAILURE;
+            }
+            if (dir[dir.length() - 1] != '/') {
+                dir += '/';
+            }
+            jsonnet_jpath_add(vm, dir.c_str());
+        } else if (arg == "-V" || arg == "--ext-str") {
+            std::string var, val;
+            if (!get_var_val(next_arg(i, args), var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_var(vm, var.c_str(), val.c_str());
+        } else if (arg == "-E" || arg == "--var" || arg == "--env") {
+            // TODO(dcunnin): Delete this in a future release.
+            std::cerr << "WARNING: jsonnet eval -E, --var and --env are deprecated,"
+                        << " please use -V or --ext-str." << std::endl;
+            std::string var, val;
+            if (!get_var_val(next_arg(i, args), var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_var(vm, var.c_str(), val.c_str());
+        } else if (arg == "--ext-str-file") {
+            std::string var, val;
+            if (!get_var_file(next_arg(i, args), "importstr", var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "-F" || arg == "--file") {
+            // TODO(dcunnin): Delete this in a future release.
+            std::cerr << "WARNING: jsonnet eval -F and --file are deprecated,"
+                        << " please use --ext-str-file." << std::endl;
+            std::string var, val;
+            if (!get_var_file(next_arg(i, args), "importstr", var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "--ext-code") {
+            std::string var, val;
+            if (!get_var_val(next_arg(i, args), var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "--code-var" || arg == "--code-env") {
+            // TODO(dcunnin): Delete this in a future release.
+            std::cerr << "WARNING: jsonnet eval --code-var and --code-env are deprecated,"
+                        << " please use --ext-code." << std::endl;
+            std::string var, val;
+            if (!get_var_val(next_arg(i, args), var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "--ext-code-file") {
+            std::string var, val;
+            if (!get_var_file(next_arg(i, args), "import", var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "--code-file") {
+            // TODO(dcunnin): Delete this in a future release.
+            std::cerr << "WARNING: jsonnet eval --code-file is deprecated,"
+                        << " please use --ext-code-file." << std::endl;
+            std::string var, val;
+            if (!get_var_file(next_arg(i, args), "import", var, val))
+                return ARG_FAILURE;
+            jsonnet_ext_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "-A" || arg == "--tla-str") {
+            std::string var, val;
+            if (!get_var_val(next_arg(i, args), var, val))
+                return ARG_FAILURE;
+            jsonnet_tla_var(vm, var.c_str(), val.c_str());
+        } else if (arg == "--tla-str-file") {
+            std::string var, val;
+            if (!get_var_file(next_arg(i, args), "importstr", var, val))
+                return ARG_FAILURE;
+            jsonnet_tla_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "--tla-code") {
+            std::string var, val;
+            if (!get_var_val(next_arg(i, args), var, val))
+                return ARG_FAILURE;
+            jsonnet_tla_code(vm, var.c_str(), val.c_str());
+        } else if (arg == "--tla-code-file") {
+            std::string var, val;
+            if (!get_var_file(next_arg(i, args), "import", var, val))
+                return ARG_FAILURE;
+            jsonnet_tla_code(vm, var.c_str(), val.c_str());
 
+        } else if (arg == "--gc-min-objects") {
+            long l = strtol_check(next_arg(i, args));
+            if (l < 0) {
+                std::cerr << "ERROR: invalid --gc-min-objects value: " << l << std::endl;
+                return ARG_FAILURE;
+            }
+            jsonnet_gc_min_objects(vm, l);
+        } else if (arg == "-t" || arg == "--max-trace") {
+            long l = strtol_check(next_arg(i, args));
+            if (l < 0) {
+                std::cerr << "ERROR: invalid --max-trace value: " << l << std::endl;
+                return ARG_FAILURE;
+            }
+            jsonnet_max_trace(vm, l);
+        } else if (arg == "--gc-growth-trigger") {
+            const char *arg = next_arg(i, args).c_str();
+            char *ep;
+            double v = std::strtod(arg, &ep);
+            if (*ep != '\0' || *arg == '\0') {
+                std::cerr << "ERROR: invalid number \"" << arg << "\"" << std::endl;
+                return ARG_FAILURE;
+            }
+            if (v < 0) {
+                std::cerr << "ERROR: invalid --gc-growth-trigger \"" << arg << "\""
+                            << std::endl;
+                return ARG_FAILURE;
+            }
+            jsonnet_gc_growth_trigger(vm, v);
+        } else if (arg == "-m" || arg == "--multi") {
+            config->evalMulti = true;
+            std::string output_dir = next_arg(i, args);
+            if (output_dir.length() == 0) {
+                std::cerr << "ERROR: -m argument was empty string" << std::endl;
+                return ARG_FAILURE;
+            }
+            if (output_dir[output_dir.length() - 1] != '/') {
+                output_dir += '/';
+            }
+            config->evalMultiOutputDir = output_dir;
+        } else if (arg == "-y" || arg == "--yaml-stream") {
+            config->evalStream = true;
+        } else if (arg == "-S" || arg == "--string") {
+            jsonnet_string_output(vm, 1);
+        } else if (arg.length() > 1 && arg[0] == '-') {
+            std::cerr << "ERROR: unrecognized argument: " << arg << std::endl;
+            return ARG_FAILURE;
         } else {
-            assert(config->cmd == FMT);
-            if (arg == "-i" || arg == "--in-place") {
-                config->fmtInPlace = true;
-            } else if (arg == "--test") {
-                config->fmtTest = true;
-            } else if (arg == "-n" || arg == "--indent") {
-                long l = strtol_check(next_arg(i, args));
-                if (l < 0) {
-                    std::cerr << "ERROR: invalid --indent value: " << l << std::endl;
-                    return ARG_FAILURE;
-                }
-                jsonnet_fmt_indent(vm, l);
-            } else if (arg == "--max-blank-lines") {
-                long l = strtol_check(next_arg(i, args));
-                if (l < 0) {
-                    std::cerr << "ERROR: invalid --max-blank-lines value: " << l << ""
-                              << std::endl;
-                    return ARG_FAILURE;
-                }
-                jsonnet_fmt_max_blank_lines(vm, l);
-            } else if (arg == "--comment-style") {
-                const std::string val = next_arg(i, args);
-                if (val == "h") {
-                    jsonnet_fmt_comment(vm, 'h');
-                } else if (val == "s") {
-                    jsonnet_fmt_comment(vm, 's');
-                } else if (val == "l") {
-                    jsonnet_fmt_comment(vm, 'l');
-                } else {
-                    std::cerr << "ERROR: invalid --comment-style value: " << val
-                              << std::endl;
-                    return ARG_FAILURE;
-                }
-            } else if (arg == "--string-style") {
-                const std::string val = next_arg(i, args);
-                if (val == "d") {
-                    jsonnet_fmt_string(vm, 'd');
-                } else if (val == "s") {
-                    jsonnet_fmt_string(vm, 's');
-                } else if (val == "l") {
-                    jsonnet_fmt_string(vm, 'l');
-                } else {
-                    std::cerr << "ERROR: invalid --string-style value: " << val
-                              << std::endl;
-                    return ARG_FAILURE;
-                }
-            } else if (arg == "--pad-arrays") {
-                jsonnet_fmt_pad_arrays(vm, true);
-            } else if (arg == "--no-pad-arrays") {
-                jsonnet_fmt_pad_arrays(vm, false);
-            } else if (arg == "--pad-objects") {
-                jsonnet_fmt_pad_objects(vm, true);
-            } else if (arg == "--no-pad-objects") {
-                jsonnet_fmt_pad_objects(vm, false);
-            } else if (arg == "--pretty-field-names") {
-                jsonnet_fmt_pretty_field_names(vm, true);
-            } else if (arg == "--no-pretty-field-names") {
-                jsonnet_fmt_pretty_field_names(vm, false);
-            } else if (arg == "--sort-imports") {
-                jsonnet_fmt_sort_imports(vm, true);
-            } else if (arg == "--no-sort-imports") {
-                jsonnet_fmt_sort_imports(vm, false);
-            } else if (arg == "--debug-desugaring") {
-                jsonnet_fmt_debug_desugaring(vm, true);
-            } else if (arg.length() > 1 && arg[0] == '-') {
-                std::cerr << "ERROR: unrecognized argument: " << arg << std::endl;
-                return ARG_FAILURE;
-            } else {
-                remaining_args.push_back(args[i]);
-            }
+            remaining_args.push_back(args[i]);
         }
     }
 
@@ -505,66 +338,13 @@ static ArgStatus process_args(int argc, const char **argv, JsonnetConfig *config
         return ARG_FAILURE;
     }
 
-    bool multiple_files_allowed = config->cmd == FMT && (config->fmtTest || config->fmtInPlace);
-    if (!multiple_files_allowed) {
-        if (remaining_args.size() > 1) {
-            std::string filename = remaining_args[0];
-            std::cerr << "ERROR: only one " << want << " is allowed\n" << std::endl;
-            return ARG_FAILURE;
-        }
+    if (remaining_args.size() > 1) {
+        std::string filename = remaining_args[0];
+        std::cerr << "ERROR: only one " << want << " is allowed\n" << std::endl;
+        return ARG_FAILURE;
     }
     config->inputFiles = remaining_args;
     return ARG_CONTINUE;
-}
-
-/** Reads from the input file or stdin into the input buffer. */
-static bool read_input_content(std::string filename, std::string *input)
-{
-    // Input file "-" tells Jsonnet to read stdin.
-    if (filename == "-") {
-        input->assign(std::istreambuf_iterator<char>(std::cin), std::istreambuf_iterator<char>());
-    } else {
-        std::ifstream f;
-        f.open(filename);
-        if (!f.good()) {
-            std::string msg = "Opening input file: " + filename;
-            perror(msg.c_str());
-            return false;
-        }
-        input->assign(std::istreambuf_iterator<char>(f), std::istreambuf_iterator<char>());
-        if (!f.good()) {
-            std::string msg = "Reading input file: " + filename;
-            perror(msg.c_str());
-            return false;
-        }
-    }
-    return true;
-}
-
-void change_special_filename(JsonnetConfig *config, std::string *filename)
-{
-    if (config->filenameIsCode) {
-        *filename = "<cmdline>";
-    } else if (*filename == "-") {
-        *filename = "<stdin>";
-    }
-}
-
-/** Gets Jsonnet code from any source into the input buffer and changes
- * the filename if it's not an actual filename (e.g. "-"). */
-static bool read_input(JsonnetConfig *config, std::string *filename, std::string *input)
-{
-    bool ok;
-    if (config->filenameIsCode) {
-        *input = *filename;
-        ok = true;
-    } else {
-        ok = read_input_content(*filename, input);
-    }
-    // Let's change the filename to something we can show the user
-    // if it is not a real filename.
-    change_special_filename(config, filename);
-    return ok;
 }
 
 /** Writes output files for multiple file output */
@@ -704,41 +484,6 @@ static bool write_output_stream(JsonnetVm *vm, char *output, const std::string &
     return true;
 }
 
-/** Writes the output JSON to the specified output file for single-file
- * output
- */
-static bool write_output_file(const char *output, const std::string &output_file)
-{
-    std::ostream *o;
-    std::ofstream f;
-
-    if (output_file.empty()) {
-        o = &std::cout;
-    } else {
-        f.open(output_file.c_str());
-        if (!f.good()) {
-            std::string msg = "Writing to output file: " + output_file;
-            perror(msg.c_str());
-            return false;
-        }
-        o = &f;
-    }
-
-    (*o) << output;
-
-    if (output_file.empty()) {
-        std::cout.flush();
-    } else {
-        f.close();
-        if (!f.good()) {
-            std::string msg = "Writing to output file: " + output_file;
-            perror(msg.c_str());
-            return false;
-        }
-    }
-    return true;
-}
-
 int main(int argc, const char **argv)
 {
     try {
@@ -764,140 +509,52 @@ int main(int argc, const char **argv)
         // Evaluate input Jsonnet and handle any errors from Jsonnet VM.
         int error;
         char *output;
-        switch (config.cmd) {
-            case EVAL: {
-                assert(config.inputFiles.size() == 1);
+        assert(config.inputFiles.size() == 1);
 
-                // Read input file.
-                std::string input;
-                if (!read_input(&config, &config.inputFiles[0], &input)) {
-                    jsonnet_destroy(vm);
-                    return EXIT_FAILURE;
-                }
+        // Read input file.
+        std::string input;
+        if (!read_input(config.filenameIsCode, &config.inputFiles[0], &input)) {
+            jsonnet_destroy(vm);
+            return EXIT_FAILURE;
+        }
 
-                if (config.evalMulti) {
-                    output = jsonnet_evaluate_snippet_multi(
-                        vm, config.inputFiles[0].c_str(), input.c_str(), &error);
-                } else if (config.evalStream) {
-                    output = jsonnet_evaluate_snippet_stream(
-                        vm, config.inputFiles[0].c_str(), input.c_str(), &error);
-                } else {
-                    output = jsonnet_evaluate_snippet(
-                        vm, config.inputFiles[0].c_str(), input.c_str(), &error);
-                }
+        if (config.evalMulti) {
+            output = jsonnet_evaluate_snippet_multi(
+                vm, config.inputFiles[0].c_str(), input.c_str(), &error);
+        } else if (config.evalStream) {
+            output = jsonnet_evaluate_snippet_stream(
+                vm, config.inputFiles[0].c_str(), input.c_str(), &error);
+        } else {
+            output = jsonnet_evaluate_snippet(
+                vm, config.inputFiles[0].c_str(), input.c_str(), &error);
+        }
 
-                if (error) {
-                    std::cerr << output;
-                    jsonnet_realloc(vm, output, 0);
-                    jsonnet_destroy(vm);
-                    return EXIT_FAILURE;
-                }
+        if (error) {
+            std::cerr << output;
+            jsonnet_realloc(vm, output, 0);
+            jsonnet_destroy(vm);
+            return EXIT_FAILURE;
+        }
 
-                // Write output JSON.
-                if (config.evalMulti) {
-                    if (!write_multi_output_files(
-                            vm, output, config.evalMultiOutputDir, config.outputFile)) {
-                        jsonnet_destroy(vm);
-                        return EXIT_FAILURE;
-                    }
-                } else if (config.evalStream) {
-                    if (!write_output_stream(vm, output, config.outputFile)) {
-                        jsonnet_destroy(vm);
-                        return EXIT_FAILURE;
-                    }
-                } else {
-                    bool successful = write_output_file(output, config.outputFile);
-                    jsonnet_realloc(vm, output, 0);
-                    if (!successful) {
-                        jsonnet_destroy(vm);
-                        return EXIT_FAILURE;
-                    }
-                }
-            } break;
-
-            case FMT: {
-                std::string output_file = config.outputFile;
-
-                if (config.fmtInPlace || config.fmtTest) {
-                    assert(config.inputFiles.size() >= 1);
-                    for (std::string &inputFile : config.inputFiles) {
-                        if (config.fmtInPlace) {
-                            output_file = inputFile;
-
-                            if (inputFile == "-") {
-                                std::cerr << "ERROR: cannot use --in-place with stdin" << std::endl;
-                                jsonnet_destroy(vm);
-                                return EXIT_FAILURE;
-                            }
-                            if (config.filenameIsCode) {
-                                std::cerr << "ERROR: cannot use --in-place with --exec"
-                                          << std::endl;
-                                jsonnet_destroy(vm);
-                                return EXIT_FAILURE;
-                            }
-                        }
-
-                        std::string input;
-                        if (!read_input(&config, &inputFile, &input)) {
-                            jsonnet_destroy(vm);
-                            return EXIT_FAILURE;
-                        }
-
-                        output = jsonnet_fmt_snippet(vm, inputFile.c_str(), input.c_str(), &error);
-
-                        if (error) {
-                            std::cerr << output;
-                            jsonnet_realloc(vm, output, 0);
-                            jsonnet_destroy(vm);
-                            return EXIT_FAILURE;
-                        }
-
-                        if (config.fmtTest) {
-                            // Check the output matches the input.
-                            bool ok = output == input;
-                            jsonnet_realloc(vm, output, 0);
-                            if (!ok) {
-                                jsonnet_destroy(vm);
-                                return 2;
-                            }
-                        } else {
-                            // Write output Jsonnet.
-                            bool successful = write_output_file(output, output_file);
-                            jsonnet_realloc(vm, output, 0);
-                            if (!successful) {
-                                jsonnet_destroy(vm);
-                                return EXIT_FAILURE;
-                            }
-                        }
-                    }
-                } else {
-                    assert(config.inputFiles.size() == 1);
-                    // Read input file.
-                    std::string input;
-                    if (!read_input(&config, &config.inputFiles[0], &input)) {
-                        jsonnet_destroy(vm);
-                        return EXIT_FAILURE;
-                    }
-
-                    output = jsonnet_fmt_snippet(
-                        vm, config.inputFiles[0].c_str(), input.c_str(), &error);
-
-                    if (error) {
-                        std::cerr << output;
-                        jsonnet_realloc(vm, output, 0);
-                        jsonnet_destroy(vm);
-                        return EXIT_FAILURE;
-                    }
-
-                    // Write output Jsonnet.
-                    bool successful = write_output_file(output, output_file);
-                    jsonnet_realloc(vm, output, 0);
-                    if (!successful) {
-                        jsonnet_destroy(vm);
-                        return EXIT_FAILURE;
-                    }
-                }
-            } break;
+        // Write output JSON.
+        if (config.evalMulti) {
+            if (!write_multi_output_files(
+                    vm, output, config.evalMultiOutputDir, config.outputFile)) {
+                jsonnet_destroy(vm);
+                return EXIT_FAILURE;
+            }
+        } else if (config.evalStream) {
+            if (!write_output_stream(vm, output, config.outputFile)) {
+                jsonnet_destroy(vm);
+                return EXIT_FAILURE;
+            }
+        } else {
+            bool successful = write_output_file(output, config.outputFile);
+            jsonnet_realloc(vm, output, 0);
+            if (!successful) {
+                jsonnet_destroy(vm);
+                return EXIT_FAILURE;
+            }
         }
 
         jsonnet_destroy(vm);
