+++ doc/learning/tools.html
@@ -76,7 +76,7 @@ title: Tooling
         The formatter is a command-line utility bundled with the C++ build of Jsonnet.  You can
         reformat files in place as so:
       </p>
-      <pre>jsonnet fmt -i *.jsonnet</pre>
+      <pre>jsonnetfmt -i *.jsonnet</pre>
       <p>
         It will manage indenting and horizontal spacing within a line and remove excess vertical
         spacing.  It will convert string literals, comments, fields and some operators into
