--- test_cmd/exec2.golden.stdout
@@ -1,5 +1,0 @@
-{
-   "a": 1,
-   "b": 2,
-   "c": 3
-}
