+++ test_cmd/fmt1.golden.stdout
@@ -1,0 +1,1 @@
+{ a: 1, b: 2, c: 3 }
