+++ test_cmd/fmt_bad_out.golden.stderr
@@ -1,0 +1,1 @@
+ERROR: -o argument was empty string
