+++ test_cmd/fmt_double_dash.golden.stdout
@@ -1,0 +1,1 @@
+-1
