+++ test_cmd/fmt_out.golden.custom_output
@@ -1,0 +1,1 @@
+{ a: 1, b: 2, c: 3 }
