+++ test_cmd/fmt_simple1.golden.stdout
@@ -1,0 +1,5 @@
+{
+  a: 1,
+  b: 2,
+  c: 3,
+}
