+++ test_cmd/fmt_simple4.golden.stderr
@@ -1,0 +1,1 @@
+Opening input file: nosuchfile.jsonnet: No such file or directory
