+++ test_cmd/fmt_simple5.golden.stderr
@@ -1,0 +1,1 @@
+ERROR: only one filename is allowed
