+++ test_cmd/fmt_simple_out.golden.custom_output
@@ -1,0 +1,5 @@
+{
+  a: 1,
+  b: 2,
+  c: 3,
+}
