+++ test_cmd/fmt_version1.golden.stdout
@@ -1,0 +1,1 @@
+Jsonnet reformatter v0.12.1
