+++ test_cmd/fmt_version2.golden.stdout
@@ -1,0 +1,1 @@
+Jsonnet reformatter v0.12.1
