+++ test_cmd/help.golden.stdout.cpp
@@ -1,14 +1,8 @@
 Jsonnet commandline interpreter v0.12.1
 
-General commandline:
-jsonnet [<cmd>] {<option>} { <filename> }
-Note: <cmd> defaults to "eval"
+jsonnet {<option>} <filename>
 
-The eval command:
-jsonnet eval {<option>} <filename>
-Note: Only one filename is supported
-
-Available eval options:
+Available options:
   -h / --help             This message
   -e / --exec             Treat filename as code
   -J / --jpath <dir>      Specify an additional library search dir (right-most wins)
@@ -41,27 +35,6 @@ in reverse order before the paths specified by --jpath (i.e. left-most wins)
 E.g. JSONNET_PATH=a:b jsonnet -J c -J d is equivalent to:
 JSONNET_PATH=d:c:a:b jsonnet
 jsonnet -J b -J a -J c -J d
-
-The fmt command:
-jsonnet fmt {<option>} { <filename> }
-Note: Some options do not support multiple filenames
-
-Available fmt options:
-  -h / --help             This message
-  -e / --exec             Treat filename as code
-  -o / --output-file <file> Write to the output file rather than stdout
-  -i / --in-place         Update the Jsonnet file(s) in place.
-  --test                  Exit with failure if reformatting changed the file(s).
-  -n / --indent <n>       Number of spaces to indent by (default 2, 0 means no change)
-  --max-blank-lines <n>   Max vertical spacing, 0 means no change (default 2)
-  --string-style <d|s|l>  Enforce double, single (default) quotes or 'leave'
-  --comment-style <h|s|l> # (h), // (s)(default), or 'leave'; never changes she-bang
-  --[no-]pretty-field-names Use syntax sugar for fields and indexing (on by default)
-  --[no-]pad-arrays       [ 1, 2, 3 ] instead of [1, 2, 3]
-  --[no-]pad-objects      { x: 1, y: 2 } instead of {x: 1, y: 2} (on by default)
-  --[no-]sort-imports     Sorting of imports (on by default)
-  --debug-desugaring      Unparse the desugared AST without executing it
-  --version               Print version
 
 In all cases:
 <filename> can be - (stdin)
