--- test_cmd/nosuchcommand.golden.stderr.cpp
@@ -1,1 +1,0 @@
-Opening input file: nosuchcommand: No such file or directory
