--- test_cmd/nosuchcommand.golden.stderr.golang
@@ -1,1 +1,0 @@
-Opening input file: nosuchcommand: no such file or directory
