+++ test_suite/run_fmt_tests.sh
@@ -42,8 +42,8 @@ for TEST in *.jsonnet ../examples/*.jsonnet ../examples/terraform/*.jsonnet ../b
     fi
 
     EXPECTED_EXIT_CODE=0
-    JSONNET_CMD="$VALGRIND "$JSONNET_BIN" fmt"
-    test_eval "$JSONNET_CMD" "$TEST" "$EXPECTED_EXIT_CODE" "$GOLDEN_OUTPUT" "$GOLDEN_KIND"
+    JSONNETFMT_CMD="$VALGRIND "$JSONNETFMT_BIN""
+    test_eval "$JSONNETFMT_CMD" "$TEST" "$EXPECTED_EXIT_CODE" "$GOLDEN_OUTPUT" "$GOLDEN_KIND"
 done
 
 if [ $FAILED -eq 0 ] ; then
